requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
DateTime==5.4
Deprecated==1.2.14
Flask==3.0.1
frozenlist==1.4.1
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.1.0
MarkupSafe==2.1.4
multidict==6.0.4
packaging==23.2
pillow==10.2.0
pycparser==2.21
PyGithub==2.1.1
PyJWT==2.8.0
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-pptx==0.6.23
pytz==2023.3.post1
requests==2.31.0
setuptools==69.0.3
six==1.16.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.16.0
XlsxWriter==3.1.9
yarl==1.9.4
zope.interface==6.1
